CGT000007 QMX6: Revert ENGR00224109 MX6 FEC optimize ENET_REF_CLK
On QMX6 with 6Quad and 6Dual, extremely bad performance was observed
on top of the imx_rel kernels 3.0.35_1.1.0 and 3.0.35_4.0.0, while
the performance on top of 3.0.15 was within expectations.

This change reverts the commit that was identified to cause the drop
in performance.

Note that further fixes for 4.0.0 might be needed, related to
the topic "ENET interupt to GPIO/wait mode", this change here
is not related to that issue.

NOTE:
The bad performance was not observed on 6DualLite and
6Solo. DualLite and Solo came in silicon rev. 1.1, while
Quad and Dual came in silicon rev. 1.2.

Further investigation into the issue is recommended, developing
a deeper understanding of the signal conditioning configuration
and checking the signal quality with measuring equipment.

This reverts commit 5722518fcba2dffcabcf3c2d5e2bd6c390392699:
=====
    commit 5722518fcba2dffcabcf3c2d5e2bd6c390392699
    Author: Fugang Duan <B38611@freescale.com>
    Date:   Thu Sep 20 15:26:49 2012 +0800

    ENGR00224109 - MX6 : FEC : optimize ENET_REF_CLK PAD configuration.

    In MX6 Arik and Rigel platforms, RGMII tx_clk clock source is from
    ENET_REF_CLK pad supplied by phy. To optimize the clk signal path,
    the ENET_REF_CLK I/O must have this configuration:
    	1. Disable on-chip pull-up, pull-down, and keeper
    	2. Disable hysteresis
    	3. Speed = 100 MHz
    	4. Slew rate = fast

    The optimizition make the bias point match the optimum point, which
    can maximize design margin.

    Signed-off-by: Fugang Duan  <B38611@freescale.com>
=====

Signed-off-by: Simon Braunschmidt <sbraun@emlix.com>